break out filters into their old file, add test
